modularizes tourRoutes and userRoutes into separate files
